# Implement first draft of autograd benchmark.