--- a/benchmarks/functional_autograd_benchmark/audio_text_models.py
+++ b/benchmarks/functional_autograd_benchmark/audio_text_models.py
@@ -591,8 +591,8 @@ def get_multiheadattn(device):
                               torch.nn.Linear(embed_dim, embed_dim, bias=False))
 
     model = MultiheadAttentionContainer(nhead, in_proj,
-                                      ScaledDotProduct(),
-                                      torch.nn.Linear(embed_dim, embed_dim, bias=False))
+                                        ScaledDotProduct(),
+                                        torch.nn.Linear(embed_dim, embed_dim, bias=False))
     model.to(device)
     params, names = make_functional(model)