@zhangguanheng66 2020-07-03
> @zhangguanheng66 We can definitely add it!
> What would be a good example of how it is used? Or a simple set of representative random inputs I can use to add it here?

Here is an example. You will have to add the dependency of torchtext.
https://github.com/pytorch/text/blob/bcb9104680eb9dc978a6bbcc2b9ca46cf2bdbed9/test/data/test_modules.py#L10