@albanD 2020-07-14
@zhangguanheng66 I added the multi head attention as a standalone thing (without downstream net or loss).
I re-used the setting from the tests in torchtext for all the sizes.